refactor(core): extract createDragHandlers (PLAN §3.1.d completion)
The drag/click cluster (handlePointerMove, handlePointerDown,
handlePointerUp, handleDragSelection, handleSingleClick,
toggleShiftMultiSelection, commitShiftMultiSelection, cancelActiveDrag,
getFrozenElementAtPosition) was 396 lines of densely-coupled pointer
logic inside init(). It now lives in a single focused controller that
encapsulates the element-detection throttle state (elementDetectionState
moved inside) and exposes 6 entry points used by the pointermove /
pointerdown / pointerup / pointercancel / keyup listeners.

The keyboardSelectedElement closure flag is now consumed via a
takeKeyboardSelectedElement callback so the controller stays pure about
it.

Removed ~15 now-unused imports from index.tsx (getComponentDisplayName,
clearElementPositionCache, getElementsAtPoint, isValidGrabbableElement,
isElementConnected, getElementsInDrag, getElementAnchorRatio,
createElementBounds, createPageRectFromBounds, getTagName,
DRAG_THRESHOLD_PX, ELEMENT_DETECTION_THROTTLE_MS,
PENDING_DETECTION_STALENESS_MS, getBoundsCenter,
isPositionInsideBounds, getAutoScrollDirection, setIsShiftMultiSelecting,
performCopyWithLabel, performCopyWithPerElementLabels,
toPageCoordinates, openContextMenuOrRunPendingDefault).

index.tsx: 1546 -> 1166 (-380). 68 drag-selection + shift-multi-select +
context-menu specs green.
