+++ packages/react-grab/src/core/index.tsx
@@ -36,6 +36,7 @@ import { createCommentModeHandlers } from "./comment-mode-handlers.js";
 import { createKeydownSpamTimer } from "./keydown-spam-timer.js";
 import { createSpaceDragRepositioning } from "./space-drag-repositioning.js";
 import { createActivationKeyHandlers } from "./activation-key-handlers.js";
+import { createDragHandlers } from "./drag-handlers.js";
 import { buildPublicApi } from "./build-public-api.js";
 import { createWindowFocusListeners } from "./window-focus-listeners.js";
 import { createToolbarStateController } from "./toolbar-state-controller.js";
@@ -46,44 +47,29 @@ import {
 } from "../utils/is-keyboard-event-triggered-by-input.js";
 import { createComponentNameForElement } from "../utils/create-component-name-for-element.js";
 import {
-  getComponentDisplayName,
   checkIsNextProject,
   resolveSource,
 } from "./context.js";
 import { createNoopApi } from "./noop-api.js";
 import { createEventListenerManager } from "./events.js";
 import {
-  clearElementPositionCache,
   getElementAtPosition,
-  getElementsAtPoint,
 } from "../utils/get-element-at-position.js";
-import { isValidGrabbableElement } from "../utils/is-valid-grabbable-element.js";
-import { isElementConnected } from "../utils/is-element-connected.js";
-import { getElementsInDrag } from "../utils/get-elements-in-drag.js";
-import { getElementAnchorRatio } from "../utils/get-element-anchor-ratio.js";
-import { createElementBounds } from "../utils/create-element-bounds.js";
 import {
-  createPageRectFromBounds,
 } from "../utils/create-bounds-from-drag-rect.js";
-import { getTagName } from "../utils/get-tag-name.js";
 import {
   ARROW_KEYS,
-  DRAG_THRESHOLD_PX,
-  ELEMENT_DETECTION_THROTTLE_MS,
-  PENDING_DETECTION_STALENESS_MS,
   MODIFIER_KEYS,
   DEFAULT_KEY_HOLD_DURATION_MS,
   MIN_HOLD_FOR_ACTIVATION_AFTER_COPY_MS,
   NEXTJS_REVALIDATION_DELAY_MS,
   DEFAULT_ACTION_ID,
 } from "../constants.js";
-import { getBoundsCenter } from "../utils/get-bounds-center.js";
 import { isCLikeKey } from "../utils/is-c-like-key.js";
 import { isTargetKeyCombination } from "../utils/is-target-key-combination.js";
 import { parseActivationKey } from "../utils/parse-activation-key.js";
 import { isEventFromOverlay } from "../utils/is-event-from-overlay.js";
 import type {
-  Position,
   Options,
   ReactGrabAPI,
   ContextMenuActionContext,
@@ -95,12 +81,11 @@ import type {
 import { DEFAULT_THEME } from "./theme.js";
 import { createPluginRegistry } from "./plugin-registry.js";
 import { getRequiredModifiers } from "./keyboard-handlers.js";
-import { createAutoScroller, getAutoScrollDirection } from "./auto-scroll.js";
+import { createAutoScroller } from "./auto-scroll.js";
 import { logIntro } from "./log-intro.js";
 import { getScriptOptions } from "../utils/get-script-options.js";
 import { isEnterCode } from "../utils/is-enter-code.js";
 import { isMac } from "../utils/is-mac.js";
-import { isPositionInsideBounds } from "../utils/is-position-inside-bounds.js";
 import { copyPlugin } from "./plugins/copy.js";
 import { commentPlugin } from "./plugins/comment.js";
 import { openPlugin } from "./plugins/open.js";
@@ -228,7 +213,6 @@ export const init = (rawOptions?: Options): ReactGrabAPI => {
 
     const shiftMultiSelect = createShiftMultiSelectState();
     const isShiftMultiSelecting = shiftMultiSelect.isActive;
-    const setIsShiftMultiSelecting = shiftMultiSelect.setActive;
     const stopShiftMultiSelecting = shiftMultiSelect.stop;
 
 
@@ -260,12 +244,6 @@ export const init = (rawOptions?: Options): ReactGrabAPI => {
       actions.setCopyStart({ x: positionX, y: positionY }, element);
     };
 
-    const elementDetectionState = {
-      lastDetectionTimestamp: 0,
-      pendingDetectionScheduledAt: 0,
-      latestPointerX: 0,
-      latestPointerY: 0,
-    };
     const dragPreviewDebounce = createDragPreviewDebounce();
     const debouncedDragPointer = dragPreviewDebounce.pointer;
     const scheduleDragPreviewUpdate = dragPreviewDebounce.schedule;
@@ -304,10 +282,7 @@ export const init = (rawOptions?: Options): ReactGrabAPI => {
       copyFeedbackCooldown,
       deactivateRenderer: () => deactivateRenderer(),
     });
-    const {
-      performCopyWithLabel,
-      performCopyWithPerElementLabels,
-    } = copyOrchestrator;
+
 
     createOverlayEffects({
       grab,
@@ -334,7 +309,6 @@ export const init = (rawOptions?: Options): ReactGrabAPI => {
       shiftSelectionLabelMouseX,
     } = dragSelectors;
 
-    const toPageCoordinates = toPageCoordinatesUtil;
     const calculateDragDistance = (endX: number, endY: number) =>
       calculateDragDistanceUtil(store.dragStart, endX, endY);
     const calculateDragRectangle = (endX: number, endY: number) =>
@@ -467,7 +441,6 @@ export const init = (rawOptions?: Options): ReactGrabAPI => {
     });
     const {
       openContextMenu,
-      openContextMenuOrRunPendingDefault,
       handleContextMenuDismiss,
       handleShowContextMenuInstance,
       handleToggleToolbarMenu,
@@ -489,396 +462,43 @@ export const init = (rawOptions?: Options): ReactGrabAPI => {
         setPendingContextMenuSelect: setIsPendingContextMenuSelect,
       });
 
-    const handlePointerMove = (clientX: number, clientY: number, isShiftHeld: boolean) => {
-      const shouldTrackPendingShiftSelection =
-        isShiftHeld &&
-        isShiftMultiSelecting() &&
-        !isDragging() &&
-        !store.pendingCommentMode &&
-        !isPendingContextMenuSelect();
-
-      if (
-        !isEnabled() ||
-        isPromptMode() ||
-        (isFrozenPhase() && !shouldTrackPendingShiftSelection) ||
-        isSelectionInteractionLocked() ||
-        store.contextMenuPosition !== null
-      ) {
-        return;
-      }
-
-      actions.setPointer({ x: clientX, y: clientY });
-
-      elementDetectionState.latestPointerX = clientX;
-      elementDetectionState.latestPointerY = clientY;
-
-      if (shouldTrackPendingShiftSelection) {
-        const candidate = getElementAtPosition(clientX, clientY);
-        if (candidate !== store.detectedElement) {
-          actions.setDetectedElement(candidate);
-        }
-        return;
-      }
-
-      const now = performance.now();
-      const isDetectionPending =
-        elementDetectionState.pendingDetectionScheduledAt > 0 &&
-        now - elementDetectionState.pendingDetectionScheduledAt < PENDING_DETECTION_STALENESS_MS;
-      if (
-        now - elementDetectionState.lastDetectionTimestamp >= ELEMENT_DETECTION_THROTTLE_MS &&
-        !isDetectionPending
-      ) {
-        elementDetectionState.lastDetectionTimestamp = now;
-        elementDetectionState.pendingDetectionScheduledAt = now;
-        setTimeout(() => {
-          const candidate = getElementAtPosition(
-            elementDetectionState.latestPointerX,
-            elementDetectionState.latestPointerY,
-          );
-          if (candidate !== store.detectedElement) {
-            actions.setDetectedElement(candidate);
-          }
-          elementDetectionState.pendingDetectionScheduledAt = 0;
-        });
-      }
-
-      if (isDragging()) {
-        if (isDragRepositioning()) {
-          const { pageX, pageY } = toPageCoordinates(clientX, clientY);
-          const delta = spaceDragRepositioning.applyPointerDelta(pageX, pageY);
-          if (delta) actions.shiftDragStart(delta);
-        }
-
-        scheduleDragPreviewUpdate(clientX, clientY);
-
-        const direction = getAutoScrollDirection(clientX, clientY);
-        const isNearEdge = direction.top || direction.bottom || direction.left || direction.right;
-
-        if (isNearEdge && !autoScroller.isActive()) {
-          autoScroller.start();
-        } else if (!isNearEdge && autoScroller.isActive()) {
-          autoScroller.stop();
-        }
-      }
-    };
-
-    const handlePointerDown = (clientX: number, clientY: number, isShiftHeld: boolean) => {
-      if (!isRendererActive() || isSelectionInteractionLocked()) return false;
-
-      if (!isShiftHeld && isShiftMultiSelecting()) {
-        stopShiftMultiSelecting();
-      }
-
-      actions.startDrag({ x: clientX, y: clientY }, isShiftHeld);
-      actions.setPointer({ x: clientX, y: clientY });
-      document.body.style.userSelect = "none";
-
-      scheduleDragPreviewUpdate(clientX, clientY);
-
-      pluginRegistry.hooks.onDragStart(clientX + window.scrollX, clientY + window.scrollY);
-
-      return true;
-    };
-
-    const toggleShiftMultiSelection = (element: Element, pointer: Position) => {
-      const wasElementSelected = store.frozenElements.includes(element);
-      const isFirstFrozenElement = store.frozenElements.length === 0;
-
-      if (!wasElementSelected) {
-        const bounds = createElementBounds(element);
-        const anchorRatio = getElementAnchorRatio(bounds, pointer);
-        shiftMultiSelect.setAnchorRatio(element, anchorRatio);
-        if (isFirstFrozenElement) {
-          const componentName = getComponentDisplayName(element) ?? undefined;
-          setResolvedComponentName(componentName);
-        }
-      }
-
-      actions.toggleFrozenElement(element);
-      clearElementPositionCache();
-      const isElementStillSelected = store.frozenElements.includes(element);
-
-      if (!isElementStillSelected) {
-        shiftMultiSelect.deleteAnchor(element);
-      }
-
-      if (store.frozenElements.length === 0) {
-        stopShiftMultiSelecting();
-        actions.unfreeze();
-        return;
-      }
-
-      // Animation freeze must run on the combined accumulated set, not just
-      // on the toggled element. freezeAllAnimations unfreezes its previous
-      // input before freezing its new input, so passing only [element] would
-      // resume animations on every previously shift-clicked element.
-      freezeAllAnimations(store.frozenElements);
-      setIsShiftMultiSelecting(true);
-      actions.setPointer(pointer);
-      // After toggleFrozenElement, the most recently changed element is
-      // either added (still in frozenElements) or removed. Anchor
-      // lastGrabbed to a still-selected element rather than to one that
-      // was just deselected.
-      actions.setLastGrabbed(
-        isElementStillSelected ? element : store.frozenElements[store.frozenElements.length - 1],
-      );
-      actions.freeze();
-      clearArrowNavigation();
-    };
-
-    const commitShiftMultiSelection = () => {
-      const accumulatedElements = store.frozenElements.filter(isElementConnected);
-
-      const perElementLabelEntries = accumulatedElements.map((element) => {
-        const tagName = getTagName(element) || "element";
-        const componentName = getComponentDisplayName(element) ?? undefined;
-        const anchorRatio = shiftMultiSelect.getAnchorRatio(element);
-        const bounds = createElementBounds(element);
-        const mouseX =
-          anchorRatio === undefined
-            ? bounds.x + bounds.width / 2
-            : bounds.x + bounds.width * anchorRatio;
-        return { element, tagName, componentName, mouseX };
-      });
-
-      stopShiftMultiSelecting();
-
-      if (accumulatedElements.length === 0) {
-        actions.unfreeze();
-        return;
-      }
-
-      if (accumulatedElements.length === 1) {
-        performCopyWithLabel({
-          element: accumulatedElements[0],
-          cursorX: perElementLabelEntries[0].mouseX,
-          selectedElements: accumulatedElements,
-          shouldDeactivateAfter: store.wasActivatedByToggle,
-        });
-        return;
-      }
-
-      performCopyWithPerElementLabels({
-        elements: accumulatedElements,
-        labelEntries: perElementLabelEntries,
-        shouldDeactivateAfter: store.wasActivatedByToggle,
-      });
-    };
-
-    const handleDragSelection = (
-      dragSelectionRect: ReturnType<typeof calculateDragRectangle>,
-      hasModifierKeyHeld: boolean,
-      isShiftHeld: boolean,
-    ) => {
-      const elements = getElementsInDrag(dragSelectionRect, isValidGrabbableElement);
-      const selectedElements =
-        elements.length > 0
-          ? elements
-          : getElementsInDrag(dragSelectionRect, isValidGrabbableElement, false);
-
-      if (selectedElements.length === 0) return;
-
-      const isShiftAccumulating =
-        isShiftHeld && !store.pendingCommentMode && !isPendingContextMenuSelect();
-
-      // In the shift-accumulating branch we must freeze on the COMBINED set
-      // (prior accumulated + newly dragged), because freezeAllAnimations
-      // unfreezes its prior input via finishAnimations() — which permanently
-      // advances WAAPI animations on previously selected elements past the
-      // freeze point. Calling it once with [...prior, ...new] keeps prior
-      // animations paused.
-      if (isShiftAccumulating) {
-        actions.addFrozenElements(selectedElements);
-      }
-      freezeAllAnimations(isShiftAccumulating ? store.frozenElements : selectedElements);
-
-      pluginRegistry.hooks.onDragEnd(selectedElements, dragSelectionRect);
-
-      if (isShiftAccumulating) {
-        const lastElement = selectedElements[selectedElements.length - 1];
-        setIsShiftMultiSelecting(true);
-        clearElementPositionCache();
-        actions.setPointer(getBoundsCenter(createElementBounds(lastElement)));
-        actions.setLastGrabbed(lastElement);
-        actions.freeze();
-        clearArrowNavigation();
-        return;
-      }
-
-      const firstElement = selectedElements[0];
-      const center = getBoundsCenter(createElementBounds(firstElement));
-
-      actions.setPointer(center);
-      actions.setFrozenElements(selectedElements);
-      const dragRect = createPageRectFromBounds(dragSelectionRect);
-      actions.setFrozenDragRect(dragRect);
-      actions.freeze();
-      actions.setLastGrabbed(firstElement);
-
-      if (store.pendingCommentMode) {
-        enterCommentModeForElement(firstElement, center.x, center.y);
-        return;
-      }
-
-      if (isPendingContextMenuSelect()) {
-        setIsPendingContextMenuSelect(false);
-        openContextMenuOrRunPendingDefault(firstElement, center);
-        return;
-      }
-
-      const shouldDeactivateAfter = store.wasActivatedByToggle && !hasModifierKeyHeld;
-
-      performCopyWithLabel({
-        element: firstElement,
-        cursorX: center.x,
-        selectedElements,
-        shouldDeactivateAfter,
-        dragRect,
-      });
-    };
-
-    const getFrozenElementAtPosition = (position: Position): Element | null => {
-      for (const element of store.frozenElements) {
-        if (!isElementConnected(element)) continue;
-        if (isPositionInsideBounds(position, createElementBounds(element))) {
-          return element;
-        }
-      }
-      return null;
-    };
-
-    const handleSingleClick = (
-      clientX: number,
-      clientY: number,
-      hasModifierKeyHeld: boolean,
-      isShiftHeld: boolean,
-    ) => {
-      const validFrozenElement = isElementConnected(store.frozenElement)
-        ? store.frozenElement
-        : null;
-
-      const validKeyboardSelectedElement = isElementConnected(keyboardSelectedElement)
-        ? keyboardSelectedElement
-        : null;
-
-      const elementAtPointer =
-        getElementsAtPoint(clientX, clientY).find(isValidGrabbableElement) ?? null;
-      const selectedElementUnderPointer =
-        elementAtPointer ??
-        (isElementConnected(store.detectedElement) ? store.detectedElement : null);
-      const selectedElement =
-        selectedElementUnderPointer ?? validFrozenElement ?? validKeyboardSelectedElement;
-      if (!selectedElement) return;
-
-      // While Shift is held we only operate on the live elementAtPointer.
-      // Falling through to the non-shift path would let the
-      // selectedElement fallback chain resolve to the previously-frozen
-      // element and fire an unintended single-element copy that races
-      // with the eventual commitShiftMultiSelection on Shift release. So
-      // we always return when Shift is held: toggle when an element is
-      // under the pointer, no-op when it isn't.
-      if (isShiftHeld && !store.pendingCommentMode && !isPendingContextMenuSelect()) {
-        if (elementAtPointer !== null) {
-          toggleShiftMultiSelection(elementAtPointer, { x: clientX, y: clientY });
-        }
-        return;
-      }
-
-      let positionX: number;
-      let positionY: number;
-
-      const didResolveFromFrozenElement =
-        selectedElementUnderPointer === null && validFrozenElement === selectedElement;
-      const didResolveFromKeyboardElement =
-        selectedElementUnderPointer === null &&
-        validFrozenElement === null &&
-        validKeyboardSelectedElement === selectedElement;
-
-      if (didResolveFromFrozenElement) {
-        positionX = pointer().x;
-        positionY = pointer().y;
-      } else if (didResolveFromKeyboardElement) {
-        const elementCenter = getBoundsCenter(createElementBounds(selectedElement));
-        positionX = elementCenter.x;
-        positionY = elementCenter.y;
-      } else {
-        positionX = clientX;
-        positionY = clientY;
-      }
-
-      keyboardSelectedElement = null;
-
-      if (store.pendingCommentMode) {
-        enterCommentModeForElement(selectedElement, positionX, positionY);
-        return;
-      }
-
-      if (isPendingContextMenuSelect()) {
-        setIsPendingContextMenuSelect(false);
-        const { wasIntercepted } = pluginRegistry.hooks.onElementSelect(selectedElement);
-        if (wasIntercepted) return;
-
-        freezeAllAnimations([selectedElement]);
-        actions.setFrozenElement(selectedElement);
-        const position = { x: positionX, y: positionY };
-        actions.setPointer(position);
-        actions.freeze();
-        openContextMenuOrRunPendingDefault(selectedElement, position);
-        return;
-      }
-
-      const shouldDeactivateAfter = store.wasActivatedByToggle && !hasModifierKeyHeld;
-
-      actions.setLastGrabbed(selectedElement);
-
-      performCopyWithLabel({
-        element: selectedElement,
-        cursorX: positionX,
-        shouldDeactivateAfter,
-      });
-    };
-
-    const cancelActiveDrag = () => {
-      if (!isDragging()) return;
-      stopSpaceDragRepositioning();
-      actions.cancelDrag();
-      autoScroller.stop();
-      document.body.style.userSelect = "";
-    };
-
-    const handlePointerUp = (
-      clientX: number,
-      clientY: number,
-      hasModifierKeyHeld: boolean,
-      isShiftHeld: boolean,
-    ) => {
-      if (!isDragging()) return;
-
-      dragPreviewDebounce.cancel();
-
-      const dragDistance = calculateDragDistance(clientX, clientY);
-      const wasDragGesture =
-        dragDistance.x > DRAG_THRESHOLD_PX || dragDistance.y > DRAG_THRESHOLD_PX;
-
-      // The rectangle needs to be calculated before endDrag() because endDrag
-      // resets dragStart in the store, which would zero out the rectangle.
-      const dragSelectionRect = wasDragGesture ? calculateDragRectangle(clientX, clientY) : null;
-
-      if (wasDragGesture) {
-        actions.endDrag();
-      } else {
-        actions.cancelDrag();
-      }
-      stopSpaceDragRepositioning();
-      autoScroller.stop();
-      document.body.style.userSelect = "";
-
-      if (dragSelectionRect) {
-        handleDragSelection(dragSelectionRect, hasModifierKeyHeld, isShiftHeld);
-      } else {
-        handleSingleClick(clientX, clientY, hasModifierKeyHeld, isShiftHeld);
-      }
-    };
+    const dragHandlers = createDragHandlers({
+      grab,
+      pluginRegistry,
+      phase,
+      elementSelectors,
+      autoScroller,
+      shiftMultiSelect,
+      spaceDragRepositioning,
+      copyOrchestrator,
+      menuHandlers,
+      commentModeHandlers: { handleToggleActive, enterCommentModeForElement, handleComment },
+      dragPreviewDebounce,
+      pointer,
+      isEnabled,
+      isPendingContextMenuSelect,
+      setIsPendingContextMenuSelect,
+      isDragRepositioning,
+      takeKeyboardSelectedElement: () => {
+        const element = keyboardSelectedElement;
+        keyboardSelectedElement = null;
+        return element;
+      },
+      scheduleDragPreviewUpdate,
+      setResolvedComponentName,
+      clearArrowNavigation,
+      toPageCoordinates: toPageCoordinatesUtil,
+      calculateDragDistance,
+      calculateDragRectangle,
+    });
+    const {
+      handlePointerMove,
+      handlePointerDown,
+      handlePointerUp,
+      commitShiftMultiSelection,
+      cancelActiveDrag,
+      getFrozenElementAtPosition,
+    } = dragHandlers;
 
     const eventListenerManager = createEventListenerManager();
